+++ GreenshotImageEditor/UnmanagedHelpers/Enumerations.cs
@@ -27,7 +27,7 @@ namespace GreenshotPlugin.UnmanagedHelpers
     /// Window Style Flags
     /// </summary>
     [Flags]
-    public enum WindowStyleFlags : uint
+    public enum WindowStyleFlags : long
     {
         //WS_OVERLAPPED       = 0x00000000,
         WS_POPUP = 0x80000000,
